Debounce the render, update for latest release
This will be the last commit as this project gets migrated to the
primary diffHTML mono-repository.

Debouncing the render in this application helps with benchmarks and
generally makes sense for performance wins by batching duplicative work
during a single render loop.

I fixed broken logic with keys by further improving them to be used in
all operations instead of by DOM position which can be incorrect and
unnecessarily tied logic to the DOM.